feat(theme-default): extract LayoutFrame component
