+++ packages/@vuepress/theme-default/src/client/components/LayoutFrame.vue
@@ -1,0 +1,117 @@
+<template>
+  <div
+    class="theme-container"
+    :class="containerClass"
+    @touchstart="onTouchStart"
+    @touchend="onTouchEnd"
+  >
+    <Navbar v-if="shouldShowNavbar" @toggle-sidebar="toggleSidebar">
+      <template #before>
+        <slot name="navbar-before" />
+      </template>
+      <template #after>
+        <slot name="navbar-after" />
+      </template>
+    </Navbar>
+
+    <div class="sidebar-mask" @click="toggleSidebar(false)" />
+
+    <Sidebar>
+      <template #top>
+        <slot name="sidebar-top" />
+      </template>
+      <template #bottom>
+        <slot name="sidebar-bottom" />
+      </template>
+    </Sidebar>
+
+    <slot name="main" />
+  </div>
+</template>
+
+<script lang="ts">
+import {
+  computed,
+  defineComponent,
+  onMounted,
+  onUnmounted,
+  ref,
+} from 'vue'
+import { useRouter } from 'vue-router'
+import { usePageFrontmatter } from '@vuepress/client'
+import type { DefaultThemePageFrontmatter } from '../../shared'
+import { useSidebarItems, useThemeLocaleData } from '../composables'
+import Navbar from './Navbar.vue'
+import Sidebar from './Sidebar.vue'
+
+export default defineComponent({
+  name: 'Layout',
+
+  components: { Navbar, Sidebar },
+
+  setup() {
+    const frontmatter = usePageFrontmatter<DefaultThemePageFrontmatter>()
+    const themeLocale = useThemeLocaleData()
+
+    // navbar
+    const shouldShowNavbar = computed(
+      () =>
+        frontmatter.value.navbar !== false && themeLocale.value.navbar !== false
+    )
+
+    // sidebar
+    const sidebarItems = useSidebarItems()
+    const isSidebarOpen = ref(false)
+    const toggleSidebar = (to?: boolean): void => {
+      isSidebarOpen.value = typeof to === 'boolean' ? to : !isSidebarOpen.value
+    }
+    const touchStart = { x: 0, y: 0 }
+    const onTouchStart = (e): void => {
+      touchStart.x = e.changedTouches[0].clientX
+      touchStart.y = e.changedTouches[0].clientY
+    }
+    const onTouchEnd = (e): void => {
+      const dx = e.changedTouches[0].clientX - touchStart.x
+      const dy = e.changedTouches[0].clientY - touchStart.y
+      if (Math.abs(dx) > Math.abs(dy) && Math.abs(dx) > 40) {
+        if (dx > 0 && touchStart.x <= 80) {
+          toggleSidebar(true)
+        } else {
+          toggleSidebar(false)
+        }
+      }
+    }
+
+    // classes
+    const containerClass = computed(() => [
+      {
+        'no-navbar': !shouldShowNavbar.value,
+        'no-sidebar': !sidebarItems.value.length,
+        'sidebar-open': isSidebarOpen.value,
+      },
+      frontmatter.value.pageClass,
+    ])
+
+    // close sidebar after navigation
+    let unregisterRouterHook
+    onMounted(() => {
+      const router = useRouter()
+      unregisterRouterHook = router.afterEach(() => {
+        toggleSidebar(false)
+      })
+    })
+    onUnmounted(() => {
+      unregisterRouterHook()
+    })
+
+    return {
+      frontmatter,
+      containerClass,
+      shouldShowNavbar,
+      toggleSidebar,
+      onTouchStart,
+      onTouchEnd,
+    }
+  },
+})
+</script>
