+++ packages/@vuepress/theme-default/src/client/layouts/Layout.vue
@@ -1,72 +1,50 @@
 <template>
-  <div
-    class="theme-container"
-    :class="containerClass"
-    @touchstart="onTouchStart"
-    @touchend="onTouchEnd"
-  >
-    <Navbar v-if="shouldShowNavbar" @toggle-sidebar="toggleSidebar">
-      <template #before>
-        <slot name="navbar-before" />
-      </template>
-      <template #after>
-        <slot name="navbar-after" />
-      </template>
-    </Navbar>
+  <LayoutFrame>
+    <template #navbar-before>
+      <slot name="navbar-before" />
+    </template>
+    <template #navbar-after>
+      <slot name="navbar-after" />
+    </template>
 
-    <div class="sidebar-mask" @click="toggleSidebar(false)" />
+    <template #sidebar-top>
+      <slot name="sidebar-top" />
+    </template>
+    <template #sidebar-bottom>
+      <slot name="sidebar-bottom" />
+    </template>
 
-    <Sidebar>
-      <template #top>
-        <slot name="sidebar-top" />
-      </template>
-      <template #bottom>
-        <slot name="sidebar-bottom" />
-      </template>
-    </Sidebar>
+    <template #main>
+      <Home v-if="frontmatter.home" />
 
-    <Home v-if="frontmatter.home" />
-
-    <Transition
-      v-else
-      name="fade-slide-y"
-      mode="out-in"
-      @before-enter="onBeforeEnter"
-      @before-leave="onBeforeLeave"
-    >
-      <Page :key="page.path">
-        <template #top>
-          <slot name="page-top" />
-        </template>
-        <template #bottom>
-          <slot name="page-bottom" />
-        </template>
-      </Page>
-    </Transition>
-  </div>
+      <Transition
+        v-else
+        name="fade-slide-y"
+        mode="out-in"
+        @before-enter="onBeforeEnter"
+        @before-leave="onBeforeLeave"
+      >
+        <Page :key="page.path">
+          <template #top>
+            <slot name="page-top" />
+          </template>
+          <template #bottom>
+            <slot name="page-bottom" />
+          </template>
+        </Page>
+      </Transition>
+    </template>
+  </LayoutFrame>
 </template>
 
 <script lang="ts">
-import {
-  computed,
-  defineComponent,
-  onMounted,
-  onUnmounted,
-  ref,
-  Transition,
-} from 'vue'
-import { useRouter } from 'vue-router'
+import { defineComponent, ref, Transition } from 'vue'
 import { usePageData, usePageFrontmatter } from '@vuepress/client'
 import type { DefaultThemePageFrontmatter } from '../../shared'
+import { useScrollPromise } from '../composables'
 import Home from '../components/Home.vue'
 import Page from '../components/Page.vue'
-import Navbar from '../components/Navbar.vue'
-import Sidebar from '../components/Sidebar.vue'
-import {
-  useScrollPromise,
-  useSidebarItems,
-  useThemeLocaleData,
-} from '../composables'
+import LayoutFrame from '../components/LayoutFrame.vue'
 
 export default defineComponent({
   name: 'Layout',
@@ -74,66 +52,13 @@ export default defineComponent({
   components: {
     Home,
     Page,
-    Navbar,
-    Sidebar,
+    LayoutFrame,
     Transition,
   },
 
   setup() {
     const page = usePageData()
     const frontmatter = usePageFrontmatter<DefaultThemePageFrontmatter>()
-    const themeLocale = useThemeLocaleData()
-
-    // navbar
-    const shouldShowNavbar = computed(
-      () =>
-        frontmatter.value.navbar !== false && themeLocale.value.navbar !== false
-    )
-
-    // sidebar
-    const sidebarItems = useSidebarItems()
-    const isSidebarOpen = ref(false)
-    const toggleSidebar = (to?: boolean): void => {
-      isSidebarOpen.value = typeof to === 'boolean' ? to : !isSidebarOpen.value
-    }
-    const touchStart = { x: 0, y: 0 }
-    const onTouchStart = (e): void => {
-      touchStart.x = e.changedTouches[0].clientX
-      touchStart.y = e.changedTouches[0].clientY
-    }
-    const onTouchEnd = (e): void => {
-      const dx = e.changedTouches[0].clientX - touchStart.x
-      const dy = e.changedTouches[0].clientY - touchStart.y
-      if (Math.abs(dx) > Math.abs(dy) && Math.abs(dx) > 40) {
-        if (dx > 0 && touchStart.x <= 80) {
-          toggleSidebar(true)
-        } else {
-          toggleSidebar(false)
-        }
-      }
-    }
-
-    // classes
-    const containerClass = computed(() => [
-      {
-        'no-navbar': !shouldShowNavbar.value,
-        'no-sidebar': !sidebarItems.value.length,
-        'sidebar-open': isSidebarOpen.value,
-      },
-      frontmatter.value.pageClass,
-    ])
-
-    // close sidebar after navigation
-    let unregisterRouterHook
-    onMounted(() => {
-      const router = useRouter()
-      unregisterRouterHook = router.afterEach(() => {
-        toggleSidebar(false)
-      })
-    })
-    onUnmounted(() => {
-      unregisterRouterHook()
-    })
 
     // handle scrollBehavior with transition
     const scrollPromise = useScrollPromise()
@@ -143,11 +68,6 @@ export default defineComponent({
     return {
       frontmatter,
       page,
-      containerClass,
-      shouldShowNavbar,
-      toggleSidebar,
-      onTouchStart,
-      onTouchEnd,
       onBeforeEnter,
       onBeforeLeave,
     }
